+++ dev/com.ibm.ws.jsf.2.2_fat/test-applications/BeanValidationTests.war/src/com/ibm/ws/jsf22/fat/beanvalidation/faces40/BeanValidationBean.java
@@ -8,7 +8,7 @@
  * Contributors:
  *     IBM Corporation - initial API and implementation
  */
-package com.ibm.ws.jsf22.fat.beanvalidation;
+package com.ibm.ws.jsf22.fat.beanvalidation.faces40;
 
 import javax.faces.bean.ManagedBean;
 import javax.faces.bean.SessionScoped;
